refactor: moving editPage APIs from "src/components/EditPage/temp_data/trip_reducer.jsx" to "src/api/editPage.js"
